Conflicts:
	qmake/library/qmakeevaluator.cpp
	(cherry picked from commit 1af6dc2c8fb4d91400fddc5050166f972ae57c9a in qttools)
	src/corelib/kernel/qcore_mac_objc.mm
	src/gui/painting/qcolor.h
	src/plugins/platforms/cocoa/qcocoawindow.mm

+++ src/plugins/platforms/direct2d/qwindowsdirect2dbitmap.cpp
@@ -80,7 +80,7 @@ public:
         bitmap.Reset();
 
         D2D1_SIZE_U size = {
-            width, height
+            UINT32(width), UINT32(height)
         };
 
         HRESULT hr = deviceContext->get()->CreateBitmap(size, data, pitch,
